Review network batch: Direct Connect ports below 1 Gbps and above 10 Gbps read their only port type, WAF Classic keeps its logging edge, block counts are float64

+++ terraform/infer/infer.go
@@ -211,7 +211,7 @@ func (b *builder) node(r *eval.Resource) model.Node {
 				n.Attributes[f.Key] = true
 			} else if c, ok := countBlocks(r.Attrs, f.TerraformPath()); ok && f.Type == field.Number {
 				// A number that points at a block reads how many are written.
-				n.Attributes[f.Key] = c
+				n.Attributes[f.Key] = float64(c)
 			}
 		}
 		for _, f := range s.Assumptions() {
